buffer: split `buffer` files into multiple directories
